Add OVS-DPDK support, for issue 104
This commit adds OVS-DPDK support to Kube-OVN. User instructions are
included in a new file docs/dpdk.md

A new Dockerfile has been added to include OVS-DPDK along with OVN.
Where DPDK is required, this image is used for the ovs-ovn pod, in
place of the existing kernel-OVS “kube-ovn” image. This Dockerfile is
currently based on Fedora 32 for reasons noted as comments within the
file. It should later be possible to change this to CentOS when full
DPDK 19 support is available.

I recommend the above Dockerfile is built and tagged as
kube-ovn-dpdk:<version>, where the version corresponds to the DPDK
version used within the image (in this case 19:11) rather than the
Kube-OVN version. I recommend this as DPDK applications have a strong
dependency on DPDK version. If we force an end user to always use the
latest version, then we will likely break their DPDK app. I propose
over time we provide images for multiple DPDK versions and let the user
pick to suit their needs. I don’t see these images or Dockerfiles
requiring maintenance or support. They should be completely independent
of Kube-OVN versions and releases.

The install.sh script has been modified. It now takes a flag
--with-dpdk=<version> so the user can indicate they want to install
OVS-DPDK based on which version of DPDK. Version of DPDK required will
determine version of OVS and this will be already built into the Docker
image provided. The Kube-OVN version installed is still set at the top
of the script as the VERSION variable. This should still be the case
going forward, Kube-OVN and DPDK versions should operate independently
of each other. However, it’s something to watch. If future versions of
Kube-OVN have a strong dependency on newer versions of OVS, then the
older version of OVS used for DPDK may become an issue. We may have to
update the install script so a user wanting an older version of DPDK
has no choice but to use an older version of Kube-OVN that’s known to
be compatible. I don’t foresee this being an issue, but one to watch as
I said.

New startup and healthcheck scripts added for OVS-DPDK.