fix checkSBBindings error when hostname is not nodeName
in some case the k8s spec.nodeName is not same with a `hostname`,
then the sbctl could not get the chassis uuid, the check process
will fail here.
A workaroud solution is that try to get chassis name first from encap
table and then do the left lookup.
It is not a good solution, when we use a second nic and change the encap
ip to the second nic.

Signed-off-by: Wan Junjie <wanjunjie@bytedance.com>