Don't enqueue VPC update when DeletionTimestamp is zero
We change the check to not enqueue a VPC update when the DeletionTimestamp
is zero. Previously, the VPC update would basically always be enqueued, as
the DeletionTimestamp would be zero until it was scheduled for deletion. This
lead to a huge amount of enqueued VPC updates and therefore a lot of Kubernetes
API requests and load on the API.

The logical not was removed in #3107 which therefore introduced this bug.

Signed-off-by: Tobias Kantusch <git@kantusch.dev>