speaker: add option to set node IP to override pod IP
Sometimes the speaker pod cannot run in host network mode, causing
the default router ID and the source IP address of some routes to
be set to the pod IP, which is in a different subnet than the actual
router IP address. By introducing a new flag for the node IP, we can
override the invalid router IP address.

Signed-off-by: Navid <yaa110@gmail.com>