Support /32 tunnel-source (#5144)
There exists a use case for sourcing tunnel traffic from a /32 address assigned to the lo interface of the node. The lo interface typically comes with a localhost address (ie. 127.0.0.1). This change excludes localhost addresses from acting as the tunnel source. It also permits using a /32 address for the tunnel source if the host-tunnel-src variable is set to true (false by default).

Signed-off-by: Anthony Timmins <anthony.timmins@crowdstrike.com>

---------

Signed-off-by: Anthony Timmins <anthony.timmins@crowdstrike.com>
Signed-off-by: netdever <46330739+netdever@users.noreply.github.com>
Co-authored-by: Anthony Timmins <anthony.timmins@crowdstrike.com>