feat: create the ability to map external Parent Keys
This feature addresses bug #5995 by creating a system that would allow an
operator to extend the known Parent IDs giving kube-ovn better means to
co-exist in mixed datacenter operations. The change introduces a new
configuration option that takes a comma separated list of strings, with
glob capablities. This new option will allow operators to provide a define
a list of external IDs which will be respected by Kube-OVNs cleanup system.

Closes Bug: https://github.com/kubeovn/kube-ovn/issues/5995
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>