feat(natgw): give pod_ips to bgp speakre to prioritize IPv4 over IPv6
When running DualStack with IPv6 before IPv4 in the podCIDR, the podIP env variable will contain an IPv6, which is not a valid router ID. This causes GoBGP to crash, failing to properly serialize its own routerID inthe BGPOpen message. The NAT GW already has code to infer the routerID from the podIPs, and will prioritize IPv4s. We just didn't pass it through the correct env variable when the speaker is in a NAT gateway

Signed-off-by: SkalaNetworks <contact@skala.network>