fix(vpcnatgateway): bgp speaker would crash if no ipv4 is available (#6059)
GoBGP v4 seems to crash when serializing with an invalid RouterID, whereas on older version, it defaulted to 0.0.0.0. There's functionally no problem with multiple peers having the same RouterID when doing eBGP. If using route reflectors, it might be more of a problem. In that case, users can use the --router-id parameter. Note that this case should not occur anyway, the NAT GW doesn't support IPv6 only networks right now as it cannot do NATv6. This is purely a failover to prevent crashes.

Signed-off-by: SkalaNetworks <contact@skala.network>
Signed-off-by: donghee-park-rsk <donghee.park@rakuten.com>